Player movement and basic inputs added
New Input and Cinemachine packages imported. Custom player input controller created and added to the player in the scene. WASD, and arrow keys move, and space bar jumps. The player controller script inherits from the Player Input Controller and calls the OnMove() and OnJump() methods when their respective buttons are pressed. This can be later changed out for unity events on the Player Input component if we desire. They do the same thing. Player controller variables are Serialized to display in inspector for debugging purposes.